requirements.txt
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.20.0
websocket-client>=1.3.0
tqdm>=4.64.0 